requirements.txt
torch==2.7.1+cu118
streamlit==1.47.0
pillow==11.0.0
torchvision==0.22.1+cu118
numpy==2.3.1